Add authentication

+++ DailyFeeling/DailyFeeling/Controllers/FeelingsController.cs
@@ -1,9 +1,11 @@
 using DailyFeeling.Models;
 using DailyFeeling.Services;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 
 namespace DailyFeeling.Controllers;
 
+[Authorize]
 [ApiController]
 [Route("api/[controller]")]
 public class FeelingsController : ControllerBase
@@ -15,30 +17,42 @@ public class FeelingsController : ControllerBase
         _feelingsService = feelingsService;
     }
 
-    [HttpGet("GetTodaysFeeling/{username}")]
-    public async Task<IActionResult> GetTodaysFeeling(string username)
+    [HttpGet("GetTodaysFeeling")]
+    public async Task<IActionResult> GetTodaysFeeling()
     {
+        var username = User.FindFirst("username")?.Value;
+        if (string.IsNullOrEmpty(username))
+            return StatusCode(ApiResponse<User?>.Unauthorized().StatusCode, ApiResponse<User?>.Unauthorized());
         var response = await _feelingsService.GetTodaysFeelingAsync(username);
         return StatusCode(response.StatusCode, response);
     }
 
-    [HttpPost("CreateFeeling/{username}/{mood}")]
-    public async Task<IActionResult> CreateFeeling(string username, Mood mood)
+    [HttpPost("CreateFeeling")]
+    public async Task<IActionResult> CreateFeeling([FromBody] Mood mood)
     {
+        var username = User.FindFirst("username")?.Value;
+        if (string.IsNullOrEmpty(username))
+            return StatusCode(ApiResponse<User?>.Unauthorized().StatusCode, ApiResponse<User?>.Unauthorized());
         var response = await _feelingsService.CreateFeelingAsync(username, mood);
         return StatusCode(response.StatusCode, response);
     }
 
-    [HttpDelete("DeleteFeeling/{username}")]
-    public async Task<IActionResult> DeleteFeeling(string username)
+    [HttpDelete("DeleteFeeling")]
+    public async Task<IActionResult> DeleteFeeling()
     {
+        var username = User.FindFirst("username")?.Value;
+        if (string.IsNullOrEmpty(username))
+            return StatusCode(ApiResponse<User?>.Unauthorized().StatusCode, ApiResponse<User?>.Unauthorized());
         var response = await _feelingsService.DeleteFeelingAsync(username);
         return StatusCode(response.StatusCode, response);
     }
 
-    [HttpGet("GetFeelingHistory/{username}")]
-    public async Task<IActionResult> GetFeelingHistory(string username)
+    [HttpGet("GetFeelingHistory")]
+    public async Task<IActionResult> GetFeelingHistory()
     {
+        var username = User.FindFirst("username")?.Value;
+        if (string.IsNullOrEmpty(username))
+            return StatusCode(ApiResponse<User?>.Unauthorized().StatusCode, ApiResponse<User?>.Unauthorized());
         var response = await _feelingsService.GetFeelingHistoryAsync(username);
         return StatusCode(response.StatusCode, response);
     }
